# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flycam Filming Request Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FilmingRequestStatus = typeof FilmingRequestStatus[keyof typeof FilmingRequestStatus];


export const FilmingRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  filming: 'filming',
  completed: 'completed',
  rejected: 'rejected',
} as const;

export interface FilmingRequest {
  id: number;
  name: string;
  email: string;
  latitude: number;
  longitude: number;
  locationName: string;
  status: FilmingRequestStatus;
  /** @nullable */
  queuePosition?: number | null;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  ipAddress?: string;
  createdAt: string;
  updatedAt: string;
}

export interface FilmingRequestInput {
  /** @minLength 1 */
  name: string;
  email: string;
  latitude: number;
  longitude: number;
  /** @minLength 1 */
  locationName: string;
}

export type QueueEntryStatus = typeof QueueEntryStatus[keyof typeof QueueEntryStatus];


export const QueueEntryStatus = {
  approved: 'approved',
  filming: 'filming',
  completed: 'completed',
} as const;

export interface QueueEntry {
  id: number;
  name: string;
  locationName: string;
  latitude: number;
  longitude: number;
  status: QueueEntryStatus;
  /** @nullable */
  queuePosition: number | null;
  /** @nullable */
  scheduledAt: string | null;
  /** @nullable */
  videoUrl: string | null;
  createdAt: string;
}

export interface QueueStats {
  pending: number;
  approved: number;
  filming: number;
  completed: number;
  total: number;
}

export interface ScheduleInput {
  scheduledAt: string;
}

export interface CompleteInput {
  videoUrl: string;
}

export interface ErrorResponse {
  error: string;
}

export type AdminListRequestsParams = {
status?: AdminListRequestsStatus;
};

export type AdminListRequestsStatus = typeof AdminListRequestsStatus[keyof typeof AdminListRequestsStatus];


export const AdminListRequestsStatus = {
  pending: 'pending',
  approved: 'approved',
  filming: 'filming',
  completed: 'completed',
  rejected: 'rejected',
} as const;