# lib/api-zod/src/generated/types/filmingRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flycam Filming Request Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { FilmingRequestStatus } from './filmingRequestStatus';

export interface FilmingRequest {
  id: number;
  name: string;
  email: string;
  latitude: number;
  longitude: number;
  locationName: string;
  status: FilmingRequestStatus;
  /** @nullable */
  queuePosition?: number | null;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  ipAddress?: string;
  createdAt: string;
  updatedAt: string;
}