# lib/api-zod/src/generated/types/filmingRequestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flycam Filming Request Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FilmingRequestInput {
  /** @minLength 1 */
  name: string;
  email: string;
  latitude: number;
  longitude: number;
  /** @minLength 1 */
  locationName: string;
}