# lib/api-zod/src/generated/types/queueEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flycam Filming Request Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { QueueEntryStatus } from './queueEntryStatus';

export interface QueueEntry {
  id: number;
  name: string;
  locationName: string;
  latitude: number;
  longitude: number;
  status: QueueEntryStatus;
  /** @nullable */
  queuePosition: number | null;
  /** @nullable */
  scheduledAt: string | null;
  /** @nullable */
  videoUrl: string | null;
  createdAt: string;
}