# lib/api-zod/src/generated/types/queueStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flycam Filming Request Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface QueueStats {
  pending: number;
  approved: number;
  filming: number;
  completed: number;
  total: number;
}